examples: Move lv2v example to new folder
